refactor(DEV-14970): restructure CreateReceivables component, remove unused utils and components, and improve error handling
