+++ packages/sdk-react/src/components/receivables/InvoiceDetails/CreateReceivable/sections/EntitySection.tsx
@@ -1,4 +1,4 @@
-import type { SectionGeneralProps } from './Section.types';
+import type { SectionGeneralProps } from './types';
 import { CreateReceivablesFormProps } from '@/components/receivables/InvoiceDetails/CreateReceivable/validation';
 import { t } from '@lingui/macro';
 import { useLingui } from '@lingui/react';
