--- packages/sdk-react/src/components/receivables/InvoiceDetails/CreateReceivable/sections/Section.types.ts
@@ -1,11 +1,0 @@
-export interface SectionGeneralProps {
-  /**
-   * Describes if the form must be disabled
-   *  to prevent the user from editing it
-   */
-  disabled: boolean;
-}
-
-export interface ValidationErrorItem {
-  [key: string]: ValidationErrorItem | { message: string } | undefined;
-}
